+++ cmd/podgrouper/app/app.go
@@ -20,12 +20,14 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/healthz"
 	"sigs.k8s.io/controller-runtime/pkg/log/zap"
+	"sigs.k8s.io/controller-runtime/pkg/manager"
 	metricsserver "sigs.k8s.io/controller-runtime/pkg/metrics/server"
 	"sigs.k8s.io/controller-runtime/pkg/webhook"
 
 	"github.com/NVIDIA/KAI-scheduler/pkg/apis/scheduling/v2"
 	kubeAiSchedulerV2alpha2 "github.com/NVIDIA/KAI-scheduler/pkg/apis/scheduling/v2alpha2"
 	controllers "github.com/NVIDIA/KAI-scheduler/pkg/podgrouper"
+	pluginshub "github.com/NVIDIA/KAI-scheduler/pkg/podgrouper/podgrouper/hub"
 	// +kubebuilder:scaffold:imports
 )
 
@@ -47,10 +49,24 @@ func init() {
 	// +kubebuilder:scaffold:scheme
 }
 
-// +kubebuilder:rbac:groups=core,resources=configmaps,verbs=get;list;watch;create;update;patch;delete
-// +kubebuilder:rbac:groups=coordination.k8s.io,resources=leases,verbs=get;list;watch;create;update;patch;delete
+type App struct {
+	Mgr               manager.Manager
+	DefaultPluginsHub pluginshub.PluginsHub
+
+	configs    controllers.Configs
+	pluginsHub pluginshub.PluginsHub
+}
 
 func Run() error {
+	app, err := New()
+	if err != nil {
+		return err
+	}
+
+	return app.Run()
+}
+
+func New() (*App, error) {
 	var opts Options
 	opts.AddFlags(flag.CommandLine)
 
@@ -87,26 +103,52 @@ func Run() error {
 		// LeaderElectionReleaseOnCancel: true,
 	})
 	if err != nil {
-		return err
+		return nil, err
 	}
 
-	if err = (&controllers.PodReconciler{
-		Client: mgr.GetClient(),
-		Scheme: mgr.GetScheme(),
-	}).SetupWithManager(mgr, configs); err != nil {
+	defaultPluginsHub := pluginshub.NewDefaultPluginsHub(mgr.GetClient(), configs.SearchForLegacyPodGroups,
+		configs.KnativeGangSchedule, configs.SchedulingQueueLabelKey, configs.NodePoolLabelKey,
+		configs.DefaultPrioritiesConfigMapName, configs.DefaultPrioritiesConfigMapNamespace)
+
+	app := &App{
+		Mgr:               mgr,
+		DefaultPluginsHub: defaultPluginsHub,
+		configs:           configs,
+		pluginsHub:        nil,
+	}
+	return app, nil
+}
+
+func (app *App) RegisterPlugins(pluginsHub pluginshub.PluginsHub) {
+	app.pluginsHub = pluginsHub
+}
+
+// +kubebuilder:rbac:groups=core,resources=configmaps,verbs=get;list;watch;create;update;patch;delete
+// +kubebuilder:rbac:groups=coordination.k8s.io,resources=leases,verbs=get;list;watch;create;update;patch;delete
+
+func (app *App) Run() error {
+	pluginsHub := app.pluginsHub
+	if pluginsHub == nil {
+		pluginsHub = app.DefaultPluginsHub
+	}
+
+	if err := (&controllers.PodReconciler{
+		Client: app.Mgr.GetClient(),
+		Scheme: app.Mgr.GetScheme(),
+	}).SetupWithManager(app.Mgr, app.configs, pluginsHub); err != nil {
 		return err
 	}
 	// +kubebuilder:scaffold:builder
 
-	if err := mgr.AddHealthzCheck("healthz", healthz.Ping); err != nil {
+	if err := app.Mgr.AddHealthzCheck("healthz", healthz.Ping); err != nil {
 		return err
 	}
-	if err := mgr.AddReadyzCheck("readyz", healthz.Ping); err != nil {
+	if err := app.Mgr.AddReadyzCheck("readyz", healthz.Ping); err != nil {
 		return err
 	}
 
 	setupLog.Info("starting manager")
-	return mgr.Start(ctrl.SetupSignalHandler())
+	return app.Mgr.Start(ctrl.SetupSignalHandler())
 }
 
 func initLogger() {